refactor(settings): unify settings on core SettingsManager and eliminate Android wrapper
