+++ app/src/main/java/com/vinnovateit/latch/features/home/HomeScreen.kt
@@ -68,7 +68,7 @@ import com.vinnovateit.latch.common.util.TooltipHint
 import com.vinnovateit.latch.core.model.LiveDataPoint
 import com.vinnovateit.latch.core.model.SessionSummary
 import com.vinnovateit.latch.features.home.components.SpectrumCard
-import com.vinnovateit.latch.features.settings.manager.SettingsManager
+import com.vinnovateit.latch.core.settings.SettingsManager
 import com.vinnovateit.latch.features.wifi.background.ForegroundService
 import com.vinnovateit.latch.features.wifi.manager.ConnectionStatus
 import com.vinnovateit.latch.ui.theme.*
