+++ app/src/main/java/com/vinnovateit/latch/features/stats/PortalAccountScreen.kt
@@ -40,7 +40,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.vinnovateit.latch.features.settings.manager.SettingsManager
+import com.vinnovateit.latch.core.settings.SettingsManager
 import com.vinnovateit.latch.core.portal.PortalHistoryClient
 import com.vinnovateit.latch.platform.LatchAppGraph
 import com.vinnovateit.latch.ui.theme.LocalIsDarkTheme
