+++ app/src/main/java/com/vinnovateit/latch/features/wifi/quicksettings/LatchTileService.kt
@@ -7,7 +7,7 @@ import android.service.quicksettings.Tile
 import android.service.quicksettings.TileService
 import android.util.Log
 import com.vinnovateit.latch.features.wifi.background.ForegroundService
-import com.vinnovateit.latch.features.settings.manager.SettingsManager
+import com.vinnovateit.latch.core.settings.SettingsManager
 import com.vinnovateit.latch.platform.LatchAppGraph
 import kotlinx.coroutines.*
 
@@ -49,7 +49,7 @@ class LatchTileService : TileService() {
         // tile is tapped right after a reboot, before MainActivity/ForegroundService/
         // the widget ever ran) - setAutoLogin() below would crash on its lateinit
         // sharedPreferences without this.
-        SettingsManager.initialize(applicationContext)
+        LatchAppGraph.initialize(applicationContext)
 
         if (isProcessing) {
             Log.d(TAG, "Already processing, ignoring click")
