+++ app/src/main/java/com/vinnovateit/latch/ui/theme/Theme.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.platform.LocalView
 import androidx.core.view.WindowInsetsControllerCompat
 import androidx.core.view.WindowCompat
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.vinnovateit.latch.features.settings.manager.SettingsManager
+import com.vinnovateit.latch.core.settings.SettingsManager
 import com.materialkolor.PaletteStyle
 import com.materialkolor.dynamicColorScheme
